Add unicode random write benchmark
This commit adds a benchmark which writes random unicode characters at
random positions on the screen.

Initially I didn't expect much difference to the normal random character
write in the alternate screen buffer, but apparently some terminals
really have issues with this.

It turns out alacritty is extremely slow with this on first run, but
then it gets better when running it again. Possibly because it needs to
find the glyphs and fonts.