Load benchmarks from external executables
This is a complete rewrite of vtebench, which aims to simplify the
benchmark creation and allow for running benchmarks without relying on
any external tools like `time` or `cat`.

All benchmarks are now sourced from external files, which are spawned as
separate executables with their stdout captured for benchmarks. A
benchmark consists of a `setup` and a `benchmark` file inside of a
directory. A benchmark directory can be specified which will be walked
recursively to look for these benchmarks.

Since benchmarks are now sourced from files, a buffer is read and stored
in memory. This allows efficient writing to the PTY and capturing of
just the write call, completely bypassing file I/O or externas commands.

Currently only a single `yes`-like benchmark is supported, but the
modularity should make it simple to recover the previous functionality
and expand on top of it.

The only supported output format is a `perf`-like stdout summary.
However more output formats like CSVs and graphs are necessary.