Use average instead of median for stdout format
The median has a few problems for a good summary that might give people
the wrong idea about the results. While it is better for having a bias
toward the mode, it is of little use when the distribution is not in a
bell shape. Since some terminals use internal buffers, there can be a
big difference between individual benchmark iterations, which can lead
to a great median with a terrible average.