Add script for multiple output plots
This adds the new `detalied.sh` script which will do the same thing as
the original `gnuplot.sh` script but output each benchmark to a
different file. Since the detailed information in a single plot were not
very useful, the original script has been removed.

This should make it a bit easier to compare two different terminals if
their differences are not significant.

The gnuplot scripts have also been moved to their own directory, to
reduce the clutter in the directory root.