Initialize history before scrolling benchmarks
Since the scrolling benchmarks should be focused on the time it takes
to scroll content into history, not the time it takes to initialize the
history, 100,000 lines of history are first written to initialize
everything.

Generally since the initialization is only done once during program
execution, it should be largely irrelevant as long as it manages to do
so in a few milliseconds.