Remove missing char benchmark from defaults
The missing cells benchmark was useful for developing improvements to
the missing cell rendering mechanism, however since then it has mostly
been causing issues. This was partially addressed in 9cb3afc by reducing
the payload size, but its inconsistency has still been an issue.

To my knowledge, this benchmark tests no paths on the terminal emulation
layer that are not already covered by other benchmarks. It also relies
on a certain client-behavior (the glyphs need to be missing). Since
improvements to the renderer can actually slow down terminal benchmarks,
having this here together with other terminal benchmarks doesn't seem
like a good fit.

Since it is a useful benchmark for certain usecases, it was just moved
to the new `extra_benchmarks` directory. This still allows running these
benchmarks on demand for people that are aware of their problems.
Hopefully this should prevent frustrating benchmark sessions running
vtebench over and over again just to verify if there's actually an issue
or not.